+++ src/main/resources/static/css/style.css
@@ -167,3 +167,124 @@ h1 {
 .search-bar input::-ms-input-placeholder {
     color: white;
 }
+
+.card {
+    display: flex;
+    align-items: center;
+    background-color: #f3f3f3;
+    border-radius: 12px;
+    padding: 2rem;
+    box-shadow: 0 2px 8px rgba(0,0,0,0.05);
+    transition: background-color 0.3s;
+    text-align: left;
+    margin-bottom: 30px;
+}
+
+.card img {
+    width: 200px;
+    height: auto;
+    border-radius: 8px;
+    margin-right: 30px;
+}
+
+.card-content {
+    flex: 1;
+}
+
+.card h2 {
+    font-size: 1.8rem;
+    margin-bottom: 1rem;
+}
+
+.card p {
+    font-size: 1.1rem;
+    margin: 0.5rem 0;
+}
+.product-list {
+    display: flex;
+    flex-wrap: wrap;
+    gap: 30px;
+    justify-content: center;
+}
+
+.card {
+    width: calc(33.333% - 20px); /* 3개씩 정렬 */
+    box-sizing: border-box;
+    display: flex;
+    align-items: center;
+    padding: 1.5rem;
+    background-color: #f3f3f3;
+    border-radius: 12px;
+    box-shadow: 0 2px 8px rgba(0,0,0,0.05);
+    transition: background-color 0.3s;
+}
+
+.card img {
+    width: 120px;
+    height: auto;
+    border-radius: 8px;
+    margin-right: 20px;
+}
+
+.card-content {
+    flex: 1;
+}
+
+@media screen and (max-width: 1000px) {
+    .card {
+        width: calc(50% - 20px); /* 태블릿: 2개씩 */
+    }
+}
+
+@media screen and (max-width: 600px) {
+    .card {
+        width: 100%; /* 모바일: 1개씩 */
+        flex-direction: column;
+        text-align: center;
+    }
+
+    .card img {
+        margin: 0 0 10px 0;
+    }
+}
+
+.product-title {
+    font-size: 1.2rem;
+    font-weight: bold;
+    margin-bottom: 6px;
+
+    word-break: break-word;        /* 단어 중간에서도 줄바꿈 허용 */
+    white-space: normal;           /* 줄바꿈 허용 */
+    overflow: hidden;
+
+    display: block;                /* 기본 블록 요소 */
+    max-height: 2.5em;             /* 예: 최대 2.5~3줄까지만 보이게 하고 넘기면 잘림 */
+}
+
+/* 가격 - 고정 높이 유지 */
+.product-price {
+    font-size: 1rem;
+    margin-bottom: 6px;
+}
+
+/* 설명 - 최대 3줄까지 표시 후 ... */
+.product-description {
+    font-size: 0.95rem;
+    color: #333;
+    line-height: 1.4;
+    margin-bottom: 10px;
+
+    display: -webkit-box;
+    -webkit-line-clamp: 2;                 /* 최대 1줄로 제한 */
+    -webkit-box-orient: vertical;
+    overflow: hidden;
+    text-overflow: ellipsis;
+
+    word-break: break-word;                /* 긴 단어도 줄바꿈 허용 */
+    white-space: normal;
+    max-height: calc(1.4em * 3);           /* 3줄 높이 고정 */
+}
+/* 카드 고정 높이 유지 */
+.card {
+    height: 280px; /* 예시 높이 조정 가능 */
+}
