+++ src/main/resources/templates/fragments/layout.html
@@ -16,7 +16,7 @@
         |
         <a href="/user/signup" class="menu-button"> 회원 가입 </a>
         |
-        <a href="/order/common/search" class="menu-button"> 상품 검색 </a>
+        <a href="/order/common/home" class="menu-button"> 상품 검색 </a>
     </div>
     <hr/>
 </div>
