+++ src/main/resources/templates/index.html
@@ -114,7 +114,7 @@
                 <input type="submit" value="로그인">
             </form>
 
-            <form action="/order/common/search" method="get">
+            <form action="/order/common/home" method="get">
                 <button type="submit" class="guest-search-button">비회원 상품 검색</button>
             </form>
