+++ src/main/resources/templates/order/common/home.html
@@ -4,8 +4,8 @@
 <body>
 
 <div class="container">
-    <h6>
-        <img src="/images/minimart-title.png" alt="Mini Mart" style="width: 700px; height: 200px;">
+    <h6 style="margin-top: 0px; margin-bottom: 0px;">
+        <img src="/images/minimart-title.png" alt="Mini Mart" style="width: 600px; height: 300px;">
     </h6>
 
 <div th:replace="~{fragments/layout.html :: fragment-body-menu}"></div>
