+++ src/main/resources/templates/order/customer/product_detail.html
@@ -42,12 +42,12 @@
 <body>
 
 <h6 style="margin-top: 0px; margin-bottom: 0px;">
-    <a th:href="@{/order/common/home}">
+    <a th:href="@{/order/customer/home}">
         <img src="/images/minimart-title.png" alt="Mini Mart" style="width: 600px; height: 300px;">
     </a>
 </h6>
 
-<div th:replace="~{fragments/layout.html :: fragment-body-menu}"></div>
+<div th:replace="~{fragments/layout.html :: fragment-customer-body-menu}"></div>
 
 <div class="container">
     <!-- 이미지 영역 -->
