--- src/main/resources/templates/order/seller/main.html
@@ -1,23 +1,0 @@
-<!doctype html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8" xmlns:th="http://www.thymeleaf.org">
-    <meta name="viewport" content="width=device-width, initial-scale=1">
-    <title>메인 페이지</title>
-    <style>
-        body {
-            text-align: center;
-        }
-    </style>
-    <link rel="stylesheet" type="text/css" href="/css/style.css">
-</head>
-<body>
-    <h1>미니 마켓 판매자 페이지</h1>
-
-    <div th:replace="~{fragments/layout.html :: fragment-sellr-body-menu}"></div>
-
-    <p th:text="'환영합니다, ' + ${userName} + '님!'"></p>
-
-    <div th:replace="fragments/layout.html :: common-footer"></div>
-</body>
-</html>
