+++ src/main/resources/templates/order/seller/my_product_list.html
@@ -26,16 +26,16 @@
 </head>
 <body>
 
-<h1 style="text-align:center;">판매 상품 목록</h1>
+<h6 style="margin-top: 0px; margin-bottom: 0px; text-align:center;">
+    <a th:href="@{/order/seller/home}">
+        <img src="/images/minimart-title.png" alt="Mini Mart" style="width: 600px; height: 300px;">
+    </a>
+</h6>
 
-<div th:replace="~{fragments/layout.html :: fragment-sellr-body-menu}"></div>
-
-<!-- 메시지 알림 안 뜸 수정해야함!! -->
-<!--<script th:if="${message}">-->
-<!--    alert([[${message}]]);-->
-<!--</script>-->
+<div th:replace="~{fragments/layout.html :: fragment-seller-body-menu}"></div>
 
 <table>
+    <h3 style="text-align:center;">판매 상품 목록</h3>
     <thead>
     <tr>
         <th>이미지</th>
@@ -49,7 +49,7 @@
     <tbody>
     <tr th:each="product : ${products}">
         <td>
-            <img th:if="${product.imageUrl}" th:src="@{${product.imageUrl}}" alt="이미지"/>
+            <img th:if="${product.imageUrl}" th:src="@{'/images/' + ${product.imageUrl}}" alt="이미지"/>
             <span th:if="${product.imageUrl == null}">-</span>
         </td>
         <td th:text="${product.productName}">상품명</td>
