--- src/main/resources/templates/user/info.html
@@ -1,115 +1,0 @@
-<!doctype html>
-<html lang="ko" xmlns:th="http://www.thymeleaf.org">
-<head>
-    <meta charset="UTF-8">
-    <title>회원 정보 수정</title>
-    <style>
-        body {
-            text-align: center;
-        }
-        table {
-            width: 80%;
-            margin: auto;
-            border-collapse: collapse;
-        }
-        th, td {
-            border: 1px solid #ccc;
-            padding: 12px;
-        }
-        h1 {
-            text-align: center;
-        }
-        input[type="text"], input[type="password"] {
-            padding: 6px;
-            width: 200px;
-        }
-        button {
-            padding: 6px 10px;
-        }
-        form.inline-form {
-            display: inline-block;
-            margin-left: 10px;
-        }
-    </style>
-    <link rel="stylesheet" type="text/css" href="/css/style.css">
-</head>
-<body>
-
-<script th:if="${param.message}" th:inline="javascript">
-    alert([[${param.message}]]);
-</script>
-
-<h1>회원 정보</h1>
-
-<div th:replace="~{fragments/layout.html :: fragment-body-menu}"></div>
-
-<table>
-    <tbody>
-    <!-- 아이디 (변경 불가) -->
-    <tr>
-        <th>아이디</th>
-        <td th:text="${user.userId}" colspan="2">userId</td>
-    </tr>
-
-    <!-- 이름 (변경 불가) -->
-    <tr>
-        <th>이름</th>
-        <td th:text="${user.userName}" colspan="2">userName</td>
-    </tr>
-
-    <!-- 비밀번호 변경 -->
-    <tr>
-        <th>비밀번호</th>
-        <td>********</td>
-        <td>
-            <form method="post" th:action="@{/user/updateField}" class="inline-form">
-                <input type="hidden" name="field" value="password"/>
-                <input type="hidden" name="userId" th:value="${user.userId}"/>
-                <input type="password" name="value" placeholder="새 비밀번호"/>
-                <button type="submit">변경</button>
-            </form>
-        </td>
-    </tr>
-
-    <!-- 전화번호 변경 -->
-    <tr>
-        <th>전화번호</th>
-        <td th:text="${user.phoneNumber}">010-xxxx-xxxx</td>
-        <td>
-            <form method="post" th:action="@{/user/updateField}" class="inline-form">
-                <input type="hidden" name="field" value="phoneNumber"/>
-                <input type="hidden" name="userId" th:value="${user.userId}"/>
-                <input type="text" name="value" placeholder="010-0000-0000"/>
-                <button type="submit">변경</button>
-            </form>
-        </td>
-    </tr>
-
-    <!-- 판매자 여부 -->
-    <tr>
-        <th>판매자 여부</th>
-        <td th:text="${user.sellerYn} ? '판매자' : '일반회원'">일반회원</td>
-        <td>
-            <form method="post"
-                  th:action="@{/user/applySeller}"
-                  th:if="${!user.sellerYn}"
-                  class="inline-form">
-                <input type="hidden" name="userId" th:value="${user.userId}"/>
-                <button type="submit">판매자 신청</button>
-            </form>
-
-            <span th:if="${user.sellerYn}" style="color: gray;"> 판매자 계정 </span>
-        </td>
-    </tr>
-
-    <!-- 가입일 (표시만) -->
-    <tr>
-        <th>가입일</th>
-        <td th:text="${#temporals.format(user.regDt, 'yyyy-MM-dd HH:mm')}" colspan="2">2025-07-21</td>
-    </tr>
-    </tbody>
-</table>
-
-<div th:replace="fragments/layout.html :: common-footer"></div>
-</body>
-</html>
